requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
numpy>=2.0.0
pandas==2.2.3
propcache==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
spotipy==2.25.1
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
